% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia_related_functions.R
\name{get_projects_articles}
\alias{get_projects_articles}
\title{Get list of articles on the various Wikimedia Foundation projects with languages}
\usage{
get_projects_articles(items)
}
\arguments{
\item{items}{The list of items in a dataframe format (as output from read_items_list)}
}
\value{
A dataframe with four variables: item identifier, title of article, language and name of the Wikimedia project it belongs to
}
\description{
This function gets the list of articles on the various Wikimedia Foundation projects with languages from the Wikidata items
}

