% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia_related_functions.R
\name{get_wikipedia_articles}
\alias{get_wikipedia_articles}
\title{Get list of Wikipedia articles with languages}
\usage{
get_wikipedia_articles(items)
}
\arguments{
\item{items}{The list of items in a dataframe format (as output from read_items_list)}
}
\value{
A dataframe with three variables: item identifier, Wikipedia article title and Wikipedia article language
}
\description{
This function gets the list of Wikipedia articles with languages from the Wikidata items
}

