% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_related_functions.R
\name{link_property_class}
\alias{link_property_class}
\title{Link the list of Wikidata items with their properties to the list of classes}
\usage{
link_property_class(properties_id, classes, properties)
}
\arguments{
\item{properties_id}{The list of properties obtained from read_property_identifier}

\item{classes}{The list of classes obtained from read_property_class}

\item{properties}{The list of classes obtained from read_property_list}
}
\value{
A dataframe with three variables: Wikidata items identifier (not property identifier), property and class
}
\description{
This function links the list of Wikidata properties id (from read_property_identifier) to the list of classes (from read_property_class) and then to the list of properties (from read_property_list)
}

