% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_from_property_code_functions.R
\name{query_item_from_property_code}
\alias{query_item_from_property_code}
\title{Query Wikidata for an item associated with a property}
\usage{
query_item_from_property_code(property, code)
}
\arguments{
\item{property}{The Wikidata identifier of the property ('P____')}

\item{code}{The number that you choose to set for the property}
}
\value{
The list of Wikidata items that are linked to that specific code
}
\description{
This function query for Wikidata items that are associated with a specific property. You have to look first for the wikidata location identifier on wikidata.org ('Q_____')
}

