% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_location_functions.R
\name{query_location_1}
\alias{query_location_1}
\title{Query Wikidata using the administrative entity (1)}
\usage{
query_location_1(city_code)
}
\arguments{
\item{city_code}{The Wikidata identifier ('Q____')}
}
\value{
The list of Wikidata items that are geolocated into that administrative entity
}
\description{
This functions query for Wikidata items using the administrative entity. You have to look first for the wikidata location identifier on wikidata.org ('Q_____')
}

