% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_location_functions.R
\name{query_location_2}
\alias{query_location_2}
\title{Query Wikidata using a radius around the city (2)}
\usage{
query_location_2(city_code, radius, lang = "en")
}
\arguments{
\item{city_code}{The Wikidata identifier ('Q____')}

\item{radius}{A number that represents the radius in kilometers}

\item{lang}{Language of the name of Wikidata items. If not specified, English will be the default language}
}
\value{
The list of Wikidata items that are geolocated into that radius around the location with Wikidata identifier
}
\description{
This functions query for Wikidata items using a radius around the city. You have to look first for the wikidata location identifier on wikidata.org ('Q_____') and to set a radius in km.
}

