% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_location_functions.R
\name{query_location_3}
\alias{query_location_3}
\title{Query Wikidata using a box around the location (3)}
\usage{
query_location_3(city_code, first_corner_city_code, first_city_corner,
  second_corner_city_code, second_city_corner)
}
\arguments{
\item{city_code}{The Wikidata identifier ('Q____')}

\item{first_corner_city_code}{The Wikidata identifier of the first box location('Q____')}

\item{first_city_corner}{The corner at which the first box location is ('NorthEast', 'NorthWest', 'SouthEast', 'SouthWest')}

\item{second_corner_city_code}{The Wikidata identifier of the second box location('Q____')}

\item{second_city_corner}{The corner at which the second box location is ('NorthEast', 'NorthWest', 'SouthEast', 'SouthWest')}
}
\value{
The list of Wikidata items that are geolocated into that box around the location with Wikidata identifier
}
\description{
This functions query for Wikidata items using a box around the location. You have to look first for the wikidata location identifier on wikidata.org ('Q_____') and to set two other locations that will become the corners of the box. Look for the identifier of these two locations and set at which corner they are in the format 'NorthEast', 'NorthWest', 'SouthEast', 'SouthWest'.
}

