% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_category_functions.R
\name{query_location_category_2}
\alias{query_location_category_2}
\title{Query Wikidata for category using a radius around the city (2)}
\usage{
query_location_category_2(city_code, radius)
}
\arguments{
\item{city_code}{The Wikidata identifier ('Q____')}

\item{radius}{A number that represents the radius in kilometers}
}
\value{
The list of Wikidata items' categories that are geolocated into that radius around the location with Wikidata identifier
}
\description{
This functions query for Wikidata items's category using a radius around the city. You have to look first for the wikidata location identifier on wikidata.org ('Q_____') and to set a radius in km.
}

