% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_related_functions.R
\name{query_property_class}
\alias{query_property_class}
\title{Get list of class from the identifier of the properties}
\usage{
query_property_class(items)
}
\arguments{
\item{items}{The list of property items' identifiers in a dataframe format (as output from read_property_identifier)}
}
\value{
This functions simply performs ther query and downloads the JSON files, so you will have to launch the read_property_class to have a result
}
\description{
This function gets the list of class from Wikidata property items' identifier that has been obtained through the query_location_property functions and read through the read_property_identifier function
}

