% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryWikidataR-package.r
\docType{package}
\name{QueryWikidataR}
\alias{QueryWikidataR}
\alias{QueryWikidataR-package}
\title{QueryWikidataR.}
\description{
Use the Query Wikidata service in R
QueryWikidataR allows to query for Wikidata items with geocoordinates in three different ways, to get the list of related Wikipedia articles with languages and to get the categories ('Instance of' in Wikidata) of those items. The package now only performs limited queries related to items with geocoordinates, but its functionalities will be enhanced.
}

