% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_category_functions.R
\name{read_category_list}
\alias{read_category_list}
\title{Read the list of cateories related to items}
\usage{
read_category_list(city_code)
}
\arguments{
\item{city_code}{The Wikidata identifier ('Q____')}
}
\value{
A dataframe with the Wikidata identifier of the item and its category (there could be multiple categories for each item)
}
\description{
This function reads the list of categories that has been downloaded
}

