% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_from_property_code_functions.R
\name{read_item_from_property_code}
\alias{read_item_from_property_code}
\title{This function reads the list of items that has been downloaded}
\usage{
read_item_from_property_code(property, code)
}
\arguments{
\item{property}{The Wikidata identifier of the property ('P____')}

\item{code}{The number that you choose to set for the property}
}
\value{
A dataframe with the Wikidata label of the item, the code that you set and the identifier of the item
}
\description{
This function reads the list of items that has been downloaded
}

