% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_related_functions.R
\name{read_property_class}
\alias{read_property_class}
\title{Read the list of classes just queries from the query_property_class function}
\usage{
read_property_class(items)
}
\arguments{
\item{items}{The list of property items' identifiers in a dataframe format (as output from read_property_identifier)}
}
\value{
A dataframe with two variables: Wikidata property items' identifier and the class to which this property belongs to
}
\description{
This function reads the list of classes from Wikidata property items' identifier that has been obtained through the query_property_class function.
}

