% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_related_functions.R
\name{read_property_identifier}
\alias{read_property_identifier}
\title{Read the list of properties with its identifier}
\usage{
read_property_identifier(city_code)
}
\arguments{
\item{city_code}{The Wikidata identifier ('Q____')}
}
\value{
A dataframe with the Wikidata identifier of the property and its label
}
\description{
This function reads the list of properties that has been downloaded with each identifier (of the property). You should use this function ONLY when you need to search for classes.
}

