% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_property_functions.R
\name{read_property_list}
\alias{read_property_list}
\title{Read the list of properties related to items}
\usage{
read_property_list(city_code)
}
\arguments{
\item{city_code}{The Wikidata identifier ('Q____')}
}
\value{
A dataframe with the Wikidata identifier of the item and its property (there could be multiple categories for each item)
}
\description{
This function reads the list of categories that has been downloaded
}

